full_requirements.txt

# Compatible versions for Rasa + Nutri RAG coexistence
# Complete RAG System Requirements
# Core Framework
# Data Processing
# Data Processing (Rasa-compatible)
# Database
# HTTP & Caching
# Hybrid retrieval & reranking
# ML/NLP Stack
# No additional requirements - using stdlib sqlite3
# Optimized for CUDA/RTX 4060 GPU
# Production-Grade RAG System Requirements
# Progress & Logging
# Python 3.11+ compatible
# Testing
# Text Processing
# Utilities
# Vector Search
# YAML support
faiss-cpu==1.7.4
fastapi==0.95.2
httpx==0.24.1
huggingface-hub==0.20.3
inflect>=6.0.4
nltk==3.9.1
numpy==1.24.3
openpyxl==3.1.2
pandas==2.0.3
pandas>=2.0.0
pyarrow==12.0.1
pydantic==1.10.11
pytest-asyncio==0.21.1
pytest==7.4.2
python-dotenv==1.0.0
python-multipart==0.0.6
python-multipart==0.0.6  # For FastAPI file uploads
pyyaml==6.0.1
rank-bm25>=0.2.2
rapidfuzz>=2.13.7
regex>=2023.0.0
requests-cache==0.9.8
requests==2.31.0
scipy>=1.10.1
sentence-transformers==2.3.1
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
transformers==4.36.2
uvicorn[standard]==0.22.0
